Fix proximity sensor under low light.
If the ambient light is too low the proximity sensor always reports "near"
which causes the screen to be immediately turned off when starting a call.
It is not possible to turn the screen on again until the other end hangs up.
Change the proximity sensor config to fix this.