Replace TI ST driver with the one from http://omapzoom.org/?p=kernel/omap.git
The original driver supplied with the HTC sources crashes and reboots the
phone as soon as a userspace process writes to /dev/tihci because of a NULL
pointer dereference in hciif_send_frame(). The cause of this seems to be not
easily fixable.

The commit message of the OMAP driver is:

tty_hci driver exposes a /dev/hci_tty character device node, that intends to
emulate a generic /dev/ttyX device that would be used by the user-space
Bluetooth stacks to send/receive data to/from the WL combo-connectivity
chipsets.

The device driver has no internal logic of its own to intrepret data & all
such logic is handled by the user-space stack.